Implement camera classes and OpenCL program for pinhole camera

+++ src/Samplers/MultiJittered.cpp
@@ -74,7 +74,7 @@ void MultiJittered::generate_samples() {
 	for (int p = 0; p < num_sets; p++)
 		for (int i = 0; i < n; i++)
 			for (int j = 0; j < n; j++) {
-				int k = j + rand() % (n - j);
+				int k = j + (int) rand() % (n - j);
 				float t = samples[i * n + j + p * num_samples].x;
 				samples[i * n + j + p * num_samples].x = samples[i * n + k + p * num_samples].x;
 				samples[i * n + k + p * num_samples].x = t;
@@ -85,7 +85,7 @@ void MultiJittered::generate_samples() {
 	for (int p = 0; p < num_sets; p++)
 		for (int i = 0; i < n; i++)
 			for (int j = 0; j < n; j++) {
-				int k = j + rand() % (n - j);
+				int k = j + (int) rand() % (n - j);
 				float t = samples[j * n + i + p * num_samples].y;
 				samples[j * n + i + p * num_samples].y = samples[k * n + i + p * num_samples].y;
 				samples[k * n + i + p * num_samples].y = t;
